+++ src/widgets/qownnotesmarkdowntextedit.cpp
@@ -26,11 +26,6 @@
 
 QOwnNotesMarkdownTextEdit::QOwnNotesMarkdownTextEdit(QWidget *parent)
     : QMarkdownTextEdit(parent, false) {
-    // Prevent infinite loops of QResizeEvents in conjunction with setPaperMargins()
-    // when the widget is resized (and leave some space for line numbers too)
-    // See: https://github.com/pbek/QOwnNotes/issues/2679
-    setMinimumWidth((int) (80 / devicePixelRatioF()));
-
     // We need to set the internal variable to true, because we start with a highlighter
     _highlightingEnabled = true;
     _highlighter = nullptr;
@@ -86,6 +81,23 @@ QOwnNotesMarkdownTextEdit::QOwnNotesMarkdownTextEdit(QWidget *parent)
 
     setContextMenuPolicy(Qt::CustomContextMenu);
     connect(this, &QOwnNotesMarkdownTextEdit::customContextMenuRequested, this, &QOwnNotesMarkdownTextEdit::onContextMenu);
+}
+
+/*
+ * Prevent infinite loops of QResizeEvents in conjunction with setPaperMargins()
+ * when the widget is resized (and leave some space for line numbers too)
+ * See: https://github.com/pbek/QOwnNotes/issues/2679
+ */
+QSize QOwnNotesMarkdownTextEdit::minimumSizeHint() const
+{
+    int lineWidthLeftMargin = _lineNumArea->isLineNumAreaEnabled() ?
+           _lineNumArea->lineNumAreaWidth() : 0;
+
+    // Let the min size be the defaultMinSize + lineNumAreaWidth + paper margin
+    auto sizeHint = QMarkdownTextEdit::minimumSizeHint();
+    sizeHint.rwidth() += lineWidthLeftMargin + 10;
+
+    return sizeHint;
 }
 
 void QOwnNotesMarkdownTextEdit::onZoom(bool in)